requirements.txt
# =============================================================================
# OCCI Evaluation Framework - Dependencies
# =============================================================================
# Ontology-Constrained Causal Induction from Legal Text
# Tested with Python 3.9+
# =============================================================================

# -----------------------------------------------------------------------------
# Azure OpenAI & LLM Providers
# -----------------------------------------------------------------------------
openai>=1.12.0              # Azure OpenAI SDK (includes AzureOpenAI client)
anthropic>=0.18.0           # Anthropic Claude API (for ablation studies)
python-dotenv>=1.0.0        # Environment variable management

# -----------------------------------------------------------------------------
# Data Processing & Storage
# -----------------------------------------------------------------------------
pandas>=2.0.0               # DataFrame operations
numpy>=1.24.0               # Numerical computing
datasets>=2.16.0            # HuggingFace datasets (CUAD download)
pyarrow>=14.0.0             # Efficient data serialization

# -----------------------------------------------------------------------------
# Natural Language Processing
# -----------------------------------------------------------------------------
spacy>=3.7.0                # NLP pipeline
sentence-transformers>=2.3.0  # Semantic similarity for decoy filtering
nltk>=3.8.0                 # Tokenization utilities

# -----------------------------------------------------------------------------
# Graph Operations & Algorithms
# -----------------------------------------------------------------------------
networkx>=3.2               # DAG operations, cycle detection

# -----------------------------------------------------------------------------
# Visualization & Reporting
# -----------------------------------------------------------------------------
matplotlib>=3.8.0           # Publication-quality figures
seaborn>=0.13.0             # Statistical visualizations
tabulate>=0.9.0             # Table formatting

# -----------------------------------------------------------------------------
# Progress & Logging
# -----------------------------------------------------------------------------
tqdm>=4.66.0                # Progress bars
rich>=13.7.0                # Rich terminal output

# -----------------------------------------------------------------------------
# Statistical Analysis
# -----------------------------------------------------------------------------
scipy>=1.11.0               # Statistical tests (confidence intervals)
scikit-learn>=1.3.0         # Metrics utilities

# -----------------------------------------------------------------------------
# Development & Testing
# -----------------------------------------------------------------------------
pytest>=7.4.0               # Unit testing
pytest-cov>=4.1.0           # Coverage reporting
black>=24.1.0               # Code formatting
isort>=5.13.0               # Import sorting
mypy>=1.8.0                 # Type checking

# -----------------------------------------------------------------------------
# Optional: For Jupyter Notebook Analysis
# -----------------------------------------------------------------------------
# jupyter>=1.0.0
# ipywidgets>=8.1.0

# Optional: Azure OpenAI
azure-identity>=1.14.0
